+++ NEWS.md
@@ -3,8 +3,10 @@
 * Development version
 * Added na.rm to qcc bootstrap summary over tau-values
 * Added main and interaction effects, and comparison of their absolute magnitudes to ml_dadas and sem_dadas outputs
+* Added moderator/intercept difference estimates for dadas-functions
 * Added abs_coef_diff_test in sem_dadas and ml_dadas to enable tests for slope difference that is not against null but a different numeric value
-* Added an estimate of scaled divergence for ml_dadas and a derived estimate of component correlation
+* Added an estimate of scaled difference of slopes for ml_dadas and a derived estimate of component correlation
+* 
 
 # multid 0.7.0