+++ NEWS.md
@@ -1,12 +1,11 @@
 # multid 0.7.0.9000
 
-* Development version
 * Added na.rm to qcc bootstrap summary over tau-values
 * Added main and interaction effects, and comparison of their absolute magnitudes to ml_dadas and sem_dadas outputs
 * Added moderator/intercept difference estimates for dadas-functions
 * Added abs_coef_diff_test in sem_dadas and ml_dadas to enable tests for slope difference that is not against null but a different numeric value
 * Added an estimate of scaled difference of slopes for ml_dadas and a derived estimate of component correlation
-* 
+* Fixed URLs and output in README
 
 # multid 0.7.0